Widen blob size column

+++ activestorage/db/migrate/20170806125915_create_active_storage_tables.rb
@@ -5,7 +5,7 @@ class CreateActiveStorageTables < ActiveRecord::Migration[5.2]
       t.string   :filename,   null: false
       t.string   :content_type
       t.text     :metadata
-      t.integer  :byte_size,  null: false
+      t.bigint   :byte_size,  null: false
       t.string   :checksum,   null: false
       t.datetime :created_at, null: false
